requirements.txt
requests~=2.26.0
torch~=1.10.0
pillow~=8.1.2
typing~=3.7.4.3
gdown~=3.12.2
torchvision~=0.11.0
opencv-python~=4.5.5.62
numpy~=1.19.5
loguru~=0.5.3
uvicorn~=0.11.7
fastapi~=0.68.1
starlette~=0.14.2
pydantic~=1.7.4
click~=7.1.2
tqdm~=4.62.3
setuptools
aiofiles~=0.6.0
python-multipart~=0.0.5